requirements.txt
langchain
pymilvus
openai
openai-whisper
torch
transformers